feat: per-exchange fee model — default 20bps replacing hardcoded 52bps Kraken fee
NEW: signals/fee_model.py — 8 exchanges (Bitfinex 0bps to Kraken 65bps)
NEW: tests/test_fee_model.py — 10 tests
CHANGED: atr_levels.py, paper_trader.py, production_engine.py — use fee model
Impact: ATR TP/SL viable at 15m, breakeven exits -20bps not -52bps

+++ tests/test_paper_trader.py
@@ -233,7 +233,7 @@ class TestFeeCalculation:
 
         assert abs(actual_fees - expected_fees) < 0.001
         assert abs(actual_pnl - expected_net) < 0.001
-        assert int(t["fee_bps"]) == ROUND_TRIP_FEE_BPS
+        assert float(t["fee_bps"]) == ROUND_TRIP_FEE_BPS
 
 
 class TestStaleSignalGate:
